style : 도메인에 맞춰서 DTO class를 디렉토리 이동
