+++ src/main/java/com/example/planergram/postTravel/cotroller/PostAttractionController.java
@@ -1,6 +1,6 @@
-package com.example.planergram.travelContents.cotroller;
+package com.example.planergram.postTravel.cotroller;
 
-import com.example.planergram.DTO.PostAttractionDTO;
+import com.example.planergram.postTravel.DTO.PostAttractionDTO;
 import com.example.planergram.service.PostAttractionService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
