+++ src/main/java/com/example/planergram/postTravel/cotroller/PostRentCarController.java
@@ -1,6 +1,6 @@
 package com.example.planergram.postTravel.cotroller;
 
-import com.example.planergram.DTO.PostRentCarDTO;
+import com.example.planergram.postTravel.DTO.PostRentCarDTO;
 import com.example.planergram.service.PostRentCarService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
