+++ src/main/java/com/example/planergram/postTravel/cotroller/PostStayController.java
@@ -1,6 +1,6 @@
 package com.example.planergram.postTravel.cotroller;
 
-import com.example.planergram.DTO.PostStayDTO;
+import com.example.planergram.postTravel.DTO.PostStayDTO;
 import com.example.planergram.service.PostStayService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
