+++ src/main/java/com/example/planergram/travelContents/DTO/AttractionDTO.java
@@ -1,4 +1,4 @@
-package com.example.planergram.DTO;
+package com.example.planergram.travelContents.DTO;
 
 import lombok.*;
