+++ src/main/java/com/example/planergram/travelContents/cotroller/RentCarController.java
@@ -1,6 +1,6 @@
 package com.example.planergram.travelContents.cotroller;
 
-import com.example.planergram.DTO.RentCarDTO;
+import com.example.planergram.travelContents.DTO.RentCarDTO;
 import com.example.planergram.service.RentCarService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
