+++ src/main/java/com/example/planergram/travelContents/cotroller/StayController.java
@@ -1,6 +1,6 @@
 package com.example.planergram.travelContents.cotroller;
 
-import com.example.planergram.DTO.StayDTO;
+import com.example.planergram.travelContents.DTO.StayDTO;
 import com.example.planergram.service.StayService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
