+++ src/main/java/com/example/planergram/travelContents/cotroller/TrainController.java
@@ -1,6 +1,6 @@
 package com.example.planergram.travelContents.cotroller;
 
-import com.example.planergram.DTO.TrainDTO;
+import com.example.planergram.travelContents.DTO.TrainDTO;
 import com.example.planergram.service.TrainService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
