+++ src/main/java/com/example/planergram/user/DTO/UserDTO.java
@@ -1,4 +1,4 @@
-package com.example.planergram.DTO;
+package com.example.planergram.user.DTO;
 
 import lombok.*;
