+++ src/main/java/com/example/planergram/userLike/cotroller/RentCarLikeController.java
@@ -1,7 +1,7 @@
 package com.example.planergram.userLike.cotroller;
 
 
-import com.example.planergram.DTO.RentCarLikeDTO;
+import com.example.planergram.userLike.DTO.RentCarLikeDTO;
 import com.example.planergram.service.RentCarLikeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
